Imported changes from Hosung which resolves OPENSTACK-585, OPENSTACK-574.
Subject: OPENSTACK-585, OPENSTACK-574
From: Hosung Hwang
To: Yue Ko
Cc: Pavel Bondar
Attachments: object_manipulator.py; dns_controller.py; ipam_controller.py

The included code is to address both OPENSTACK-585 and OPENSTACK-574.

Pavel, I saw your change and integrated it here and modified further to address 574.
Feel free to review.

I don.t think you need to call delete_dns_view from delete_network since this is called by delete_subnet already which is called by delete_network.

Thanks,

Hosung

+++ neutron/neutron/ipam/drivers/infoblox/ipam_controller.py
@@ -327,7 +327,15 @@ class InfobloxIPAMController(neutron_ipam.NeutronIPAMController):
             self.delete_subnet(context, subnet)
 
         if not self.infoblox.has_networks(net_view):
-            self.infoblox.delete_network_view(net_view)
+            # no network ranges exist but dns view may exist, so check for dns view(s)
+            exists_dns_view = False
+            for dns_view in self.infoblox.get_dns_view(net_view):
+                if dns_view and dns_view['name']:
+                    exists_dns_view = True
+
+            # if no dns view, then safe to delete network view
+            if not exists_dns_view:
+                self.infoblox.delete_network_view(net_view)
 
         fixed_address_ref = self.ib_db.get_management_ip_ref(context,
                                                              network_id)
